feat(policies): add all policies
As mentioned a few times before, the outputted HTML from Termly is horrifically complicated and non-compliant. Trying to shoe horn that into the rest of our website caused it to render badly.

So we're using a separate sub folder, with its own styling that does support the horrific html. We can potentially integrate it into the regular website later, but its common for many companies to use a separate HTML architecture/site for their policies. Our goal is speed and reliability and this works and looks ok.